One-step estimator: mediation parameter under mediator-outcome confounding

One-step estimator: mediation parameter under mediator-outcome confounding

est_onestep(data, contrast, g_lrnr_stack, e_lrnr_stack, m_lrnr_stack,
  q_lrnr_stack, r_lrnr_stack, u_lrnr_stack, v_lrnr_stack, w_names, m_names,
  cv_folds = 10)

Arguments

data
A data.table containing the observed data, with columns in the order specified by the NPSEM (Y, M, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated as part of a call to the user-facing wrapper function medoutcon.

contrast
...

g_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a model for the propensity score, i.e., \(g = P(A | W)\).

e_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a cleverly parameterized propensity score that includes the mediators, i.e., \(e = P(A | Z, W)\).

m_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting the outcome regression.

q_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a regression involving the mediator-outcome confounder, i.e., \(q(L | A', W)\).

r_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a regression involving the mediator-outcome confounder, i.e., \(r(L | A', M, W)\).

u_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(u(L, A, W) = E[m(A, L, Z, W) * (q(L|A,W) / r(L|A,Z,W)) * (e(a'|Z,W) / e(A|Z,W)) | L = l, A = a, W = w]\).

v_lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(v(A,W) = E[\int_z m(a', l, Z, W) * q(l|A',W) d\nu(z) | A = a, W = w)]\).

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by a call to the wrapper function medoutcon.

m_names
A character vector of the names of the columns that correspond to mediators (M). The input for this argument is automatically generated by a call to the wrapper function medoutcon.

cv_folds
A numeric integer value specifying the number of folds to be created for cross-validation. Use of cross-validation / cross-fitting allows for entropy conditions on the one-step estimator to be relaxed. Note: for compatibility with origami::make_folds, this value specified here must be greater than or equal to 2; the default is to create 10 folds.