Fit pseudo-outcome regression conditioning on treatment and baseline

Fit pseudo-outcome regression conditioning on treatment and baseline

fit_nuisance_v(train_data, valid_data, contrast, lrnr_stack, m_out, q_out,
  m_names, w_names)

Arguments

train_data
A data.table containing observed data, with columns in the order specified by the NPSEM (Y, M, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated as part of a call to the user-facing wrapper function medoutcon.

valid_data
A holdout data set, with columns exactly matching those appearing in the preceding argument data, to be used for estimation via cross-fitting. Not optional for this nuisance parameter

contrast
A numeric double indicating the two values of the intervention A to be compared. The default value of c(0, 1) assumes a binary intervention node A, though support for categorical interventions is planned for future releases.

lrnr_stack
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a model for this nuisance parameter.

m_out
...

q_out
...

m_names
A character vector of the names of the columns that correspond to mediators (M). The input for this argument is automatically generated by a call to the wrapper function medoutcon.

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by a call to the wrapper function medoutcon.