EIF for stochastic interventional (in)direct effects

EIF for stochastic interventional (in)direct effects

cv_eif(
  fold,
  data_in,
  contrast,
  g_learners,
  h_learners,
  b_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  effect_type = c("interventional", "natural"),
  w_names,
  m_names
)

Arguments

fold
Object specifying cross-validation folds as generated by a call to make_folds.

data_in
A data.table containing the observed data with columns are in the order specified by the NPSEM (Y, M, Z, A, W), with column names set appropriately based on the input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated by medoutcon.

contrast
A numeric double indicating the two values of the intervention A to be compared. The default value of NULL has no effect, as the value of the argument effect is instead used to define the contrasts. To override effect, provide a numeric double vector, giving the values of a' and a*, e.g., c(0, 1).

g_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a model for the propensity score.

h_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a model for a parameterization of the propensity score that conditions on the mediators.

b_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a model for the outcome regression.

q_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a model for a nuisance regression of the intermediate confounder, conditioning on the treatment and potential baseline covariates.

r_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a model for a nuisance regression of the intermediate confounder, conditioning on the mediators, the treatment, and potential baseline confounders.

u_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a pseudo-outcome regression required for in the efficient influence function.

v_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing instantiated learners from sl3; used to fit a pseudo-outcome regression required for in the efficient influence function.

effect_type
A character indicating whether components of the interventional or natural (in)direct effects are to be estimated. In the case of the natural (in)direct effects, estimation of several nuisance parameters is unnecessary.

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by medoutcon.

m_names
A character vector of the names of the columns that correspond to mediators (M). The input for this argument is automatically generated by medoutcon.