One-step for stochastic (in)direct effects with intermediate confounding

One-step for stochastic (in)direct effects with intermediate confounding

est_onestep(
  data,
  contrast,
  g_learners,
  e_learners,
  m_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  w_names,
  m_names,
  ext_weights = NULL,
  cv_folds = 10
)

Arguments

data
A data.table containing the observed data, with columns in the order specified by the NPSEM (Y, M, Z, A, W), with column names set appropriately based on the original input data. Such a structure is merely a convenience utility to passing data around to the various core estimation routines and is automatically generated as part of a call to the wrapper function medoutcon.

contrast
A numeric double indicating the two values of the intervention A to be compared. The default value of NULL has no effect, as the value of the argument effect is instead used to define the contrasts. To override effect, provide a numeric double vector, giving the values of a' and a* (e.g., c(0, 1).

g_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a model for the propensity score, i.e., \(g = P(A | W)\).

e_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a cleverly parameterized propensity score that includes the mediators, i.e., \(e = P(A | Z, W)\).

m_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting the outcome regression.

q_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a regression involving the mediator-outcome confounder, i.e., \(q(Z | A', W)\).

r_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a regression involving the mediator-outcome confounder, i.e., \(r(Z | A', M, W)\).

u_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(u(Z, A, W) = E[m(A, Z, M, W) * (q(Z|A,W) / r(Z|A,M,W)) * (e(a'|M,W) / e(A|M,W)) | Z = z, A = a, W = w]\).

v_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(v(A,W) = E[\int_m m(a', Z, M, W) * q(z|A',W) d\nu(m) | A = a, W = w)]\).

w_names
A character vector of the names of the columns that correspond to baseline covariates (W). The input for this argument is automatically generated by a call to the wrapper function medoutcon.

m_names
A character vector of the names of the columns that correspond to mediators (M). The input for this argument is automatically generated by a call to the wrapper function medoutcon.

ext_weights
A numeric vector of observation-level weights that have been computed externally. Such weights are used in the construction of a re-weighted one-step estimator or in solving a re-weighted estimating equation in the case of the TML estimator. Use with caution.

cv_folds
A numeric integer specifying the number of folds to be created for cross-validation. Use of cross-validation / cross-fitting allows for entropy conditions on the one-step estimator to be relaxed. For compatibility with origami::make_folds, this value specified must be greater than or equal to 2; the default is to create 10 folds.