Efficient estimation of stochastic (in)direct effects in the presence of mediator-outcome confounders affected by exposure

Efficient estimation of stochastic (in)direct effects in the presence of mediator-outcome confounders affected by exposure

medoutcon(W, A, Z, M, Y, obs_weights = rep(1, length(Y)),
  ext_weights = NULL, effect = c("direct", "indirect"),
  contrast = NULL, g_learners = sl3::Lrnr_glm_fast$new(family =
  stats::binomial()), e_learners = sl3::Lrnr_glm_fast$new(family =
  stats::binomial()), m_learners = sl3::Lrnr_glm_fast$new(),
  q_learners = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  r_learners = sl3::Lrnr_glm_fast$new(family = stats::binomial()),
  u_learners = sl3::Lrnr_glm_fast$new(),
  v_learners = sl3::Lrnr_glm_fast$new(), estimator = c("onestep",
  "tmle", "ipw", "sub"), estimator_args = list(cv_folds = 10, max_iter =
  5))

Arguments

W
A matrix, data.frame, or similar object corresponding to a set of baseline covariates.

A
A numeric vector corresponding to a treatment variable. The parameter of interest is defined as a location shift of this quantity.

Z
A numeric vector corresponding to a mediator-outcome confounder affected by treatment (on the causal pathway between intervention A, mediator M, and outcome Y, but unaffected itself by the mediator M).

M
A numeric vector, matrix, data.frame, or similar corresponding to a set of mediators (on the causal pathway between the intervention A and the outcome Y).

Y
A numeric vector corresponding to an outcome variable.

obs_weights
A numeric vector of observation-level weights to be incorporated in all procedures estimating nuisance parameters. The default is to give all observations equal weighting.

ext_weights
A numeric vector of observation-level weights that have been computed externally. Such weights are used in the construction of a re-weighted one-step estimator or in solving a re-weighted estimating equation in the case of the TML estimator. Note that, unlike the argument obs_weights, these weights are not incorporated in the estimation of nuisance parameters. Input weights should be normalized. Use with caution.

effect
A character indicating whether to compute the direct effect or the indirect effect as discussed in CITE PAPER. Note that this is ignored when the argument contrast is provided.

contrast
A numeric double indicating the two values of the intervention A to be compared. The default value of NULL has no effect, as the value of the argument effect is instead used to define the contrasts. To override effect, provide a numeric double vector, giving the values of a' and a* (e.g., c(0, 1).

g_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a model for the propensity score, i.e., \(g = P(A | W)\).

e_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, for use in fitting a cleverly parameterized propensity score that includes the mediators, i.e., \(e = P(A | M, W)\).

m_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting the outcome regression.

q_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a regression involving the mediator-outcome confounder, i.e., \(q(Z | A, W)\).

r_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a regression involving the mediator-outcome confounder, i.e., \(r(Z | A, M, W)\).

u_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(u(Z, A, W) = E[m(A, Z, M, W) * (q(Z | A, W) / r(Z | A, M, W)) * (e(a' | M, W) / e(A | M, W)) | Z = z, A = a, W = w]\).

v_learners
A Stack object, or other learner class (inheriting from Lrnr_base), containing a single or set of instantiated learners from the sl3 package, to be used in fitting a reduced regression useful for computing the efficient one-step estimator, i.e., \(v(A, W) = E[\int_z m(a', z, M, W) * q(z | a', W) d \nu(z) | A = a, W = w)]\).

estimator
The desired estimator of the direct or indirect effect (or contrast-specific parameter) to be computed. Currently, the singular option is an efficient one-step estimator, though support is planned for classical estimators (substitution; inverse probability weighted) and an efficient targeted maximum likelihood estimator.

estimator_args
A list of extra arguments to be passed (via ...) to the function call for the specified estimator. The default is so chosen as to allow the number of folds used in computing the one-step estimator to be easily adjusted.